requirements.txt
pydantic
networkx
pandas
numpy
pytest
pyyaml
